+++ hydromt/data_catalog/drivers/__init__.py
@@ -3,7 +3,6 @@
 from hydromt.data_catalog.drivers.base_driver import BaseDriver, DriverOptions
 from hydromt.data_catalog.drivers.dataframe import DataFrameDriver, PandasDriver
 from hydromt.data_catalog.drivers.dataset import DatasetDriver, DatasetXarrayDriver
-from hydromt.data_catalog.drivers.dataset.xarray_driver import DatasetXarrayOptions
 from hydromt.data_catalog.drivers.geodataframe import (
     GeoDataFrameDriver,
     GeoDataFrameTableDriver,
@@ -17,7 +16,6 @@ from hydromt.data_catalog.drivers.geodataset import (
     GeoDatasetVectorDriver,
     GeoDatasetXarrayDriver,
 )
-from hydromt.data_catalog.drivers.geodataset.geodataset_driver import GeoDatasetOptions
 from hydromt.data_catalog.drivers.preprocessing import (
     harmonise_dims,
     remove_duplicates,
@@ -29,18 +27,16 @@ from hydromt.data_catalog.drivers.raster.raster_dataset_driver import (
 )
 from hydromt.data_catalog.drivers.raster.raster_xarray_driver import (
     RasterDatasetXarrayDriver,
-    RasterXarrayOptions,
 )
 from hydromt.data_catalog.drivers.raster.rasterio_driver import (
     RasterioDriver,
     RasterioOptions,
 )
+from hydromt.data_catalog.drivers.xarray_options import XarrayDriverOptions
 
 __all__ = [
     "BaseDriver",
-    "DatasetXarrayOptions",
     "GeoDataFrameTableOptions",
-    "GeoDatasetOptions",
     "DriverOptions",
     "DatasetDriver",
     "DatasetXarrayDriver",
@@ -54,13 +50,13 @@ __all__ = [
     "PyogrioDriver",
     "RasterDatasetDriver",
     "RasterDatasetXarrayDriver",
-    "RasterXarrayOptions",
     "RasterioDriver",
     "RasterioOptions",
     "harmonise_dims",
     "remove_duplicates",
     "round_latlon",
     "to_datetimeindex",
+    "XarrayDriverOptions",
 ]
 
 # define hydromt driver entry points